requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langgraph
langchain-community
langchain-ollama
chromadb
python-dotenv
pyodbc
SQLAlchemy
